Bugfix: The warning dialogs now work properly
The "SAS_Warning" function was insane. It had an absolutely non-sensical
check at the top, saying something like "if the user doesn't want UI
warnings, AND this function isn't called with 'force' OR the user has
disabled auto-restore warnings; then skip this warning"...

It's obvious that the intended purpose was to separate the features
so that the "no UI warnings" and "no auto-restore warnings" options
didn't conflict with each other.

But the end-result was actually just so non-sensical that it NEVER
warned about auto-restoring action bars! Because the function was
always called with "force" enabled (so the "or auto-restore warnings"
check never ran).

All of that has now been fixed. The stupid "force" parameter has been
removed, and we instead check the type of warning. If it's an auto-
restore dialog then we look for "no auto-restore warnings"; otherwise
we look for "no general UI warnings", and if both fail then we show the
warning.

This commit also fixes the help-tooltip for the "Show warnings frame"
option, since it was incorrectly claiming to only apply to
saving/loading sets, but it actually applies to ALL warnings.