Bugfix: Fix tooltip crash on empty action slots
When you hover over empty action slots in the Simple Action Sets GUI,
it could get an empty scan-result from the "SAS_ParseActionInfo"
function, which in turn caused the code to attempt to set an empty
tooltip, which failed with a crash.

That has now been fixed. When there is no action in the slot you hover
over, we hide any existing tooltip and do not attempt to set any new
game tooltip!