Bugfix: Only save sets if player provides a name
This bug was pretty funny... When you try to save a set, a popup dialog
appears and tells you to give it a name. The original author put a lot
of effort into designing that dialog, making it show a red warning and
disabling the "Okay" button until you type a name.

But all of that... could be defeated by simply pressing Enter, since he
had coded an "on enter: save" feature... haha.

As a result, you saved a set with no name, making it impossible to
select or delete that set again. Bad stuff!

Furthermore, the user could just type a bunch of spaces and save that,
which was also horrible...

All of that has now been fixed. We now always TRIM the setname (to
remove leading and trailing whitespace) and THEN check for emptiness,
to ensure that the user can never save action sets with corrupt names.