Enhance: Confirm discarding changes on close
When you press the "Cancel" button in the bottom right, you are asked
if you want to discard the changes (if you've edited the active set).

That feature has now been extended to the "X"-close button in the top
right of the window too, for improved safety.

NOTE:

The only remaining ways to close the window, which *will* discard the
changes without confirmation, are by pressing Escape, or by toggling
via the minimap button. The "escape" problem would be too annoying
to fix (the window is a UIPanel, and Blizzard are the ones implementing
the escape-close feature, and we have none-to-little control over it),
and the "minimap toggle" is no issue at all since it's very far away,
so the user would have to make a very conscious decision to click it.